✨  BT-1418 integrate foundry & hardhat + migrating solidity tests (#76)
* ✨ integrate foundry alongside hardhat for dual testing framework

- Add foundry.toml configuration
- Reorganize test structure: test/forge/ and test/hardhat/
- Add Foundry scripts to package.json (build:forge, test:forge, coverage:forge)
- Update hardhat.config.ts with evmVersion for compatibility
- Add GitHub workflows and coverage scripts from template
- Add Husky hooks (pre-commit, post-checkout, post-merge)
- Add .vscode/settings.json for better IDE support
- Disable coverage checks until tests are migrated (commented with TODO)

This migration is based on TokenySolutions/contracts-template and enables
gradual migration of tests from Hardhat (TypeScript) to Foundry (Solidity).

Related to: BT-1418

* ✅ trex factory tests

* ⚰️ remove trex suit

* feat(forge): reorganize Foundry test structure and sync GitHub workflows

- Move TREXFactorySetup to helpers folder
- Rename TREXSuite.t.sol to TREXFactory.t.sol
- Update fixture to inherit from Test for proper Foundry pattern
- Transfer ownership to deployer address to match Hardhat behavior
- Sync GitHub workflows with package.json scripts
- Update test.yml to use npm run build and npm run test
- Update push_checking.yml to use unified npm run lint
- Update forge_coverage.sh to use npm run coverage

* fix(ci): install Foundry and skip husky in push_checking workflow

- Add Foundry installation step before npm ci
- Set HUSKY=0 to skip husky setup in CI environment
- Fixes 'forge: not found' error during npm prepare script

* 🔧 keep hardhat settings

* ✅ create addressess directly

* ✅ removed test_StandardAddressesSet

* ✅ fix lint

* 💡 removed title natspecs

* ✅ migrated trex factort tests

* ✅ added test module multicall

* ✅ migrate agent role test

* ✅ migrate claim topics registry

* ✅ migrated trusted issuer registry

* ✅ migrate identity registry storage

* ✅ migrated identity registry tests

* ✅ migrated implementation authority test

* ✅ migrated token agen restrictions

* ✅ migrated token information test

* ✅ migrated token recover

* ✅ migrated token permit

* ✅ migrated token transfer

* ✅ migrated utilities upgrade

* ✅ migrated freeze check

* ✅ migrated compliance check in utilities test

* ✅ migrated eligibility check

* ✅ migrated transfer check in utilities

* ✅ migrated compliance test

* ✅ migrated gateway tests

* 🔥 removed already defined events

* 🗑️ imported Initializable error from openzeppelin

* ➕ added already defined events and errors

* ✅ added base test token

* ✅ used idFactory

* ♻️ refactored compliance test

* ⚰️ remoced useless vm.prank

* 🔧 deleted hardhat configurations

* 💚 fixed ci

* refactor: remove 100% coverage requirement from forge_coverage.sh

- Remove strict 100% coverage checks that were causing workflow failures
- Script now only generates and displays coverage report without failing
- Aligns with previous behavior where coverage was informational only

* ✅ resolve getVerifiedDetails

* ✅ applied _getEligibility and remove via-ir from foundry.toml

* ✅ cover tokena and token permit

* ✅ cover compliance

* ✅ cover identity registry

* ✅ increased trex gateway coverage

* ✅ increase coverage for trex factory

* ✅ increased AbstractModuleUpgradeable.sol coverage

* ⚰️ removed abstract module

* 🔧 updted generate coverage

* ✅ coverage 100% identity registry

* ✅ coverage 100% trex gate way

* ✅ coverage trex factory

* ✅ coverage 100% AbstractModuleUpgradeable by test module

* ✅ coverage IAFactory.sol

* ✅ coverage 100% for implementation authority

* ✅ coverage 1005 abstract proxy

* ✅ coverage 100% claim topic registry

* ✅ coverage 100% identity registry storage

* ✅ coverage 100% modular compliance proxy

* ✅ coverage trusted issuer registry proxy

* ✅ coverage 100% identity registry

* ✅ coverage 100% token proxy

* 🚨 resolve linter in token information

* ✅ coverage 100% roles

* 🚨 resolve linter

* 🔧 added no coverage for test, _testContracts and  utils

* ✅ coverage 100% for token.sol

* ✅ coverage 100% for modular compliance

* ✅ coverage 100% trusted issuer registry

* 🔧 restore package.json

* ✅ coverage 100% for utlity checker

* 🚨 fix linter

+++ contracts/_testContracts/ModuleNotPnP.sol
@@ -63,14 +63,14 @@
 pragma solidity 0.8.30;
 
 import "../compliance/modular/IModularCompliance.sol";
-import "../token/IToken.sol";
 import "../compliance/modular/modules/AbstractModuleUpgradeable.sol";
+import "../token/IToken.sol";
 
 // basic test contract showcasing the behavior of a module not plug & play
 contract ModuleNotPnP is AbstractModuleUpgradeable {
 
     /// state variables
-    mapping(address => uint) private _complianceData;
+    mapping(address => uint256) private _complianceData;
     mapping(address => bool) private _moduleReady;
 
     /// functions
@@ -83,12 +83,12 @@ contract ModuleNotPnP is AbstractModuleUpgradeable {
         __AbstractModule_init();
     }
 
-    function doSomething(uint _value) external onlyComplianceCall {
+    function doSomething(uint256 _value) external onlyComplianceCall {
         _complianceData[msg.sender] = _value;
     }
 
     function setModuleReady(address compliance, bool ready) external {
-        require(msg.sender == Ownable(compliance).owner(),"only compliance owner can call");
+        require(msg.sender == Ownable(compliance).owner(), "only compliance owner can call");
         _moduleReady[compliance] = ready;
     }
 
@@ -96,45 +96,48 @@ contract ModuleNotPnP is AbstractModuleUpgradeable {
      *  @dev See {IModule-moduleTransferAction}.
      *  no transfer action required in this module
      */
-    function moduleTransferAction(address _from, address _to, uint256 _value) external override onlyComplianceCall {
-    }
+    function moduleTransferAction(address _from, address _to, uint256 _value) external override onlyComplianceCall { }
 
     /**
      *  @dev See {IModule-moduleMintAction}.
      *  no mint action required in this module
      */
-    function moduleMintAction(address _to, uint256 _value) external override onlyComplianceCall {
-    }
+    function moduleMintAction(address _to, uint256 _value) external override onlyComplianceCall { }
 
     /**
      *  @dev See {IModule-moduleBurnAction}.
      *  no burn action required in this module
      */
-    function moduleBurnAction(address _from, uint256 _value) external override onlyComplianceCall {
-    }
+    function moduleBurnAction(address _from, uint256 _value) external override onlyComplianceCall { }
 
     /**
      *  @dev See {IModule-moduleCheck}.
      *  always returns true (just a test module)
      */
     function moduleCheck(
-        address /*_from*/,
+        address,
+        /*_from*/
         address _to,
         uint256 _value,
         address _compliance
-    ) external view override returns (bool) {
+    )
+        external
+        view
+        override
+        returns (bool)
+    {
         return true;
     }
 
     /**
-      *  @dev See {IModule-canComplianceBind}.
+     *  @dev See {IModule-canComplianceBind}.
      */
     function canComplianceBind(address _compliance) external view returns (bool) {
         return _moduleReady[_compliance];
     }
 
     /**
-      *  @dev See {IModule-isPlugAndPlay}.
+     *  @dev See {IModule-isPlugAndPlay}.
      */
     function isPlugAndPlay() external pure returns (bool) {
         return false;
@@ -146,4 +149,5 @@ contract ModuleNotPnP is AbstractModuleUpgradeable {
     function name() public pure returns (string memory _name) {
         return "ModuleNotPnP";
     }
+
 }
